fix: compute battery power in 64-bit so it survives real currents
millivolts * milliamps overflows uint32_t above ~73 A at 58.5 V, and the ESC
reports up to 500 A. At a real 50 V / 150 A the wrapped product reported
3.2 kW instead of 7.5 kW -- understating power exactly during climb-out.

Both the $XCTOD sentence XCTrack reads and the telemetry page's JSON carried
this. BleControl was about to be a third copy.

+++ src/WebServer/ControllerWebServer.cpp
@@ -737,7 +737,8 @@ void ControllerWebServer::startAP() {
         // kW x10 to avoid float over JSON transport (7 => 0.7 kW)
         uint16_t powerKwX10 = 0;
         if (isPowerKwAvailable()) {
-            const uint32_t powerMilliWatts = ((uint32_t) batteryVoltageMv * batteryCurrentMa) / 1000;
+            // 64-bit: see BleControl::fillTelemetry() for the overflow threshold.
+            const uint32_t powerMilliWatts = (uint32_t) (((uint64_t) batteryVoltageMv * batteryCurrentMa) / 1000);
             powerKwX10 = (uint16_t) (powerMilliWatts / 100000);
         }
